refactor: Update the file structure to be more organized
