+++ src/Components/Homepage.js
@@ -35,42 +35,42 @@ const Homepage = () => {
 
   return (
     <ImageBackground
-      source={require("./assets/home-hero-scaled-1.jpg")} 
+      source={require("../../assets/home-hero-scaled-1.jpg")} 
       style={styles.backgroundImage}
     >
     <View style={styles.screen}>
       <View style={styles.gridContainer}>
         <TouchableOpacity testID="timer-home" style={styles.box} onPress={() => handleBoxPress(1)}>
-          <Image source={require("./assets/timer.png")} />
+          <Image source={require("../../assets/timer.png")} />
           <Text style={styles.boxText}>Meditation Timer</Text>
         </TouchableOpacity>
 
         <TouchableOpacity testID="programs" style={styles.box} onPress={() => handleBoxPress(2)}>
-          <Image source={require("./assets/calendar.png")} style={styles.boxImage} />
+          <Image source={require("../../assets/calendar.png")} style={styles.boxImage} />
           <Text style={styles.boxText}>Programs</Text>
         </TouchableOpacity>
       </View>
 
       <View style={styles.gridContainer}>
         <TouchableOpacity testID="videos-home"style={styles.box} onPress={() => handleBoxPress(3)}>
-          <Image source={require("./assets/videos.png")} style={styles.boxImage} />
+          <Image source={require("../../assets/videos.png")} style={styles.boxImage} />
           <Text style={styles.boxText}>Videos</Text>
         </TouchableOpacity>
 
         <TouchableOpacity testID="gallery-home"style={styles.box} onPress={() => handleBoxPress(4)}>
-          <Image source={require("./assets/register.png")} style={styles.registerImage} />
+          <Image source={require("../../assets/register.png")} style={styles.registerImage} />
           <Text style={styles.boxText}>Gallery</Text>
         </TouchableOpacity>
       </View>
 
       <View style={styles.gridContainer}>
         <TouchableOpacity testID="donate-home" style={styles.box} onPress={() => handleBoxPress(5)}>
-          <Image source={require("./assets/donate.png")} style={styles.boxImage} />
+          <Image source={require("../../assets/donate.png")} style={styles.boxImage} />
           <Text style={styles.boxText}>Donate</Text>
         </TouchableOpacity>
 
         <TouchableOpacity testID="stats-home" style={styles.box} onPress={() => handleBoxPress(6)}>
-          <Image source={require("./assets/meditate.png")} style={styles.meditationImage} />
+          <Image source={require("../../assets/meditate.png")} style={styles.meditationImage} />
           <Text style={styles.boxText}>Meditation Stats</Text>
         </TouchableOpacity>
       </View>
